clean up tests, start onboarding bundler tests
