+++ packages/app/package.json
@@ -41,6 +41,7 @@
     "expo-image-picker": "~14.3.1",
     "expo-linking": "~5.0.2",
     "expo-secure-store": "~12.3.1",
+    "permissionless": "^0.0.16",
     "react-hook-form": "^7.48.2",
     "react-native-safe-area-context": "4.6.3",
     "react-native-svg": "13.9.0",
